fix(tracking): TRK01 — address codex-review P2 findings (DI order, amount scale, archive WORM, pagination, test gaps)
- Make IDocumentIntake wiring order-independent: Ingestion registers NoOpDocumentIntake via TryAddScoped (safe default), Documents keeps services.Replace; corrected misleading comments; added registration test proving Replace overrides prior registration
- Add storage-integrity guard (RequireMonetaryScale) in Document.CreateDetected rejecting monetary totals with > 2 decimals (numeric(18,2) truncation would alter audited amounts per CLAUDE.md §4); unit tests added
- Add append-only/WORM triggers (reject UPDATE/DELETE/TRUNCATE) to archive_entries (V005 migration); integration tests verify rejection; MODULE.md and INVARIANTS.md updated
- Compute pagination offset as long in PostgresDocumentQueries to prevent int overflow on extreme page numbers
- Strengthen append-only tests with assertion of rejection message; add page-2 assertion to GetByState pagination test

+++ src/Modules/Documents/Infrastructure/Queries/PostgresDocumentQueries.cs
@@ -75,7 +75,7 @@ public sealed class PostgresDocumentQueries : IDocumentQueries
     {
         var boundedPage = page < 1 ? 1 : page;
         var boundedPageSize = pageSize < 1 ? 1 : Math.Min(pageSize, MaxPageSize);
-        var offset = (boundedPage - 1) * boundedPageSize;
+        var offset = (long)(boundedPage - 1) * boundedPageSize;
 
         using var conn = await _connectionFactory.OpenAsync(cancellationToken);
